chore: folder entry store
