+++ src/renderer/src/components/entry-column/picture-item.tsx
@@ -3,7 +3,7 @@ import { SwipeImages } from "@renderer/components/ui/image/swipe-images"
 import { ReactVirtuosoItemPlaceholder } from "@renderer/components/ui/placeholder"
 import dayjs from "@renderer/lib/dayjs"
 import { cn } from "@renderer/lib/utils"
-import { useEntry } from "@renderer/store/entry"
+import { useEntry } from "@renderer/store/entry/hooks"
 
 import { usePreviewImages } from "../ui/image/hooks"
 import type { UniversalItemProps } from "./types"
